Implemented the spawn mechanism using an areaportal
It's workin A.O.K and an unintentinonal behavior is that when the player respawns the whole introduction sequence is skipped which is actually kind of good in a way. Still need to match the lighting of the tube room with the spawn, perhaps just a door should be used between the two rooms?